DatabaseConfigurationError, refactoring
