+++ tests/integration_tests/test_demos.py
@@ -12,7 +12,7 @@ import demo_quipuswap.models
 import demo_tezos_domains.models
 import demo_tezos_domains_big_map.models
 import demo_tzcolors.models
-from dipdup.utils import tortoise_wrapper
+from dipdup.utils.database import tortoise_wrapper
 
 
 class DemosTest(IsolatedAsyncioTestCase):
